// src/stores/character/types.ts
export interface CharacterStats {
  health?: number | null;
  magicka?: number | null;
  stamina?: number | null;
  healthBase?: number | null;
  magickaBase?: number | null;
  staminaBase?: number | null;
  level?: number | null;
  xp?: number | null;
  xpNext?: number | null;
  inventoryWeight?: number | null;
  carryWeight?: number | null;
  gold?: number | null;
}